backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.5.2
pydantic-settings==2.1.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
PyJWT==2.8.0
python-dotenv==1.0.0
